Refactoring some Rogue filtering code to be cleaner and more efficient.
