+++ quasar/rogue_queue_test.py
@@ -1,23 +1,10 @@
 import json
 import time
 
+
 from .config import config
 from .queue import QuasarQueue
-
-
-class RogueQueue(QuasarQueue):
-
-    def __init__(self):
-        super(RogueQueue, self).__init__(config.AMQP_URI, config.ROGUE_QUEUE,
-                                         config.QUASAR_EXCHANGE)
-
-    def process_message(self, message_data):
-        message = json.loads(message_data)
-        print(''.join(("Timestamp {} for Rogue message consumed."
-                       "")).format(message['data']['data']['timestamp']))
-
-
-rogue_queue = RogueQueue()
+from .rogue_queue import RogueQueue
 
 
 class TestQueue(QuasarQueue):
@@ -25,6 +12,7 @@ class TestQueue(QuasarQueue):
     def __init__(self):
         super(TestQueue, self).__init__(config.AMQP_URI, config.TEST_QUEUE,
                                         config.QUASAR_EXCHANGE)
+        self.rogue_queue = RogueQueue()
 
     def pub_generic_messages(self):
         for i in range(100):
@@ -47,27 +35,28 @@ class TestQueue(QuasarQueue):
         if ('message_source' in message['data']['meta'] and
                 message['data']['meta']['message_source'] == 'rogue'):
             print("Routing Rogue message to Rogue queue.")
-            rogue_queue.pub_message(message_data)
+            self.rogue_queue.pub_message(message_data)
         else:
             message_num = message['data']['meta']['message_num']
             print(''.join(("Generic message {} processed."
                            "")).format(message_num))
 
 
-test_queue = TestQueue()
-
-
 def pub_generic():
+    test_queue = TestQueue()
     test_queue.pub_generic_messages()
 
 
 def pub_rogue():
+    test_queue = TestQueue()
     test_queue.pub_test_rogue_messages()
 
 
 def test_consume():
+    test_queue = TestQueue()
     test_queue.start_consume()
 
 
 def rogue_consume():
+    rogue_queue = RogueQueue()
     rogue_queue.start_consume()
